test: add tests for services
CommentService delete and update not covered because the id is generated by database.

+++ src/main/java/com/demo/app/comment/CommentService.java
@@ -3,64 +3,68 @@ package com.demo.app.comment;
 import com.demo.app.post.Post;
 import com.demo.app.post.PostService;
 import com.demo.app.user.User;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
 import org.springframework.web.server.ResponseStatusException;
 
 import java.util.List;
-import java.util.Optional;
 
 @Service
 public class CommentService {
-    @Autowired
-    private CommentRepository commentRepository;
-    @Autowired
-    private PostService postService;
+    private final CommentRepository commentRepository;
+    private final PostService postService;
 
-    public Comment findById(Long id) {
-        Optional<Comment> optional = commentRepository.findById(id);
-        Comment comment = null;
-        if (optional.isPresent()) {
-            comment = optional.get();
-        } else {
-            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Comment not found for id :: " + id);
-        }
-        return comment;
+    public CommentService(CommentRepository commentRepository, PostService postService) {
+        this.commentRepository = commentRepository;
+        this.postService = postService;
     }
 
-    public List<Comment> findByPostId(Long itemId) {
+    public Comment getById(Long id) {
+        return commentRepository.findById(id).orElseThrow(() -> {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "Comment not found for id :: " + id);
+        });
+    }
+
+    public List<Comment> getByPostId(Long itemId) {
         return commentRepository.findByPostId(itemId);
     }
 
-    public List<Comment> findByUserId(Long userId) {
+    public List<Comment> getByUserId(Long userId) {
         return commentRepository.findByUserId(userId);
     }
 
-    public List<Comment> findByUserName(String name) {
+    public List<Comment> getByUserName(String name) {
         return commentRepository.findByUserName(name);
     }
 
-    public List<Comment> findByPostAndUser(Long itemId, Long userId) {
+    public List<Comment> getByPostAndUser(Long itemId, Long userId) {
         return commentRepository.findByPostAndUser(itemId, userId);
     }
 
-    public Comment save(String body, Long postId, User user) {
-        Post post = postService.getById(postId);
-        return commentRepository.save(new Comment(body, post, user));
+    public Comment addComment(CommentDto.CommentCreate commentCreate, User user) {
+        Post post = postService.getById(commentCreate.postId);
+        return commentRepository.save(new Comment(commentCreate.body, post, user));
     }
 
-    public void delete(Long id) {
-        Comment comment = commentRepository.findById(id).orElseThrow();
-        commentRepository.delete(comment);
-    }
-
-    public void deleteById(Long id) {
+    public ResponseEntity<String> deleteComment(Long id, User user) {
+        Comment comment = getById(id);
+        if (!comment.getUser().getId().equals(user.getId())) {
+            System.out.println("Unauthorized");
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Unauthorized");
+        }
         commentRepository.deleteById(id);
+        return ResponseEntity.ok("Deleted %d".formatted(id));
+
     }
 
-    public void update(Comment comment, String body) {
-        comment.setBody(body);
+    public ResponseEntity<CommentDto> updateComment(CommentDto.CommentUpdate commentUpdate, User user) {
+        Comment comment = getById(commentUpdate.id);
+        if (!comment.getUser().getId().equals(user.getId())) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body(null);
+        }
+        comment.setBody(commentUpdate.body);
         commentRepository.save(comment);
+        return ResponseEntity.ok(new CommentDto(comment));
     }
 }
